stream_dvdnav: cast to int64_t before multiplication
Just to silence static analysis warning. It is assigned to int64_t
anyway.

+++ stream/stream_dvdnav.c
@@ -456,8 +456,8 @@ static int control(stream_t *stream, int cmd, void *arg)
         int64_t tm = (int64_t)(d * 90000);
         if (tm < 0)
             tm = 0;
-        if (priv->duration && tm >= (priv->duration * 90))
-            tm = priv->duration * 90 - 1;
+        if (priv->duration && tm >= (int64_t)priv->duration * 90)
+            tm = (int64_t)priv->duration * 90 - 1;
         uint32_t pos, len;
         if (dvdnav_get_position(dvdnav, &pos, &len) != DVDNAV_STATUS_OK)
             break;
